Move TestRequestCreator -> TestCustomRequestCreator
