+++ Tests/ApolloTests/RequestCreatorTests.swift
@@ -11,7 +11,7 @@ import XCTest
 import StarWarsAPI
 
 class RequestCreatorTests: XCTestCase {
-  private let customRequestCreator = TestRequestCreator()
+  private let customRequestCreator = TestCustomRequestCreator()
   private let apolloRequestCreator = ApolloRequestCreator()
 
   private func checkString(_ string: String,
@@ -361,58 +361,5 @@ Alpha file content.
     let req = customRequestCreator.requestBody(for: query, sendOperationIdentifiers: false)
 
     XCTAssertEqual(query.queryDocument, req["test_query"] as? String)
-  }
-}
-
-fileprivate struct TestRequestCreator: RequestCreator {
-  public func requestBody<Operation: GraphQLOperation>(for operation: Operation, sendOperationIdentifiers: Bool) -> GraphQLMap {
-    var body: GraphQLMap = [
-      "test_variables": operation.variables,
-      "test_operationName": operation.operationName,
-    ]
-
-    if sendOperationIdentifiers {
-      guard let operationIdentifier = operation.operationIdentifier else {
-        preconditionFailure("To send operation identifiers, Apollo types must be generated with operationIdentifiers")
-      }
-
-      body["test_id"] = operationIdentifier
-    } else {
-      body["test_query"] = operation.queryDocument
-    }
-
-    return body
-  }
-
-  public func requestMultipartFormData<Operation: GraphQLOperation>(for operation: Operation,
-                                                                        files: [GraphQLFile],
-                                                                        sendOperationIdentifiers: Bool,
-                                                                        serializationFormat: JSONSerializationFormat.Type,
-                                                                        manualBoundary: String?) throws -> MultipartFormData {
-    let formData: MultipartFormData
-
-    if let boundary = manualBoundary {
-      formData = MultipartFormData(boundary: boundary)
-    } else {
-      formData = MultipartFormData()
-    }
-
-    let fields = requestBody(for: operation, sendOperationIdentifiers: false)
-    for (name, data) in fields {
-      if let data = data as? GraphQLMap {
-        let data = try serializationFormat.serialize(value: data)
-        formData.appendPart(data: data, name: name)
-      } else if let data = data as? String {
-        try formData.appendPart(string: data, name: name)
-      } else {
-        try formData.appendPart(string: data.debugDescription, name: name)
-      }
-    }
-
-    files.forEach {
-      formData.appendPart(inputStream: $0.inputStream, contentLength: $0.contentLength, name: $0.fieldName, contentType: $0.mimeType, filename: $0.originalName)
-    }
-
-    return formData
   }
 }
